hyperparamAndBugs

+++ oticon_utils/sound_data_module.py
@@ -96,7 +96,7 @@ class SoundDataModule(pl.LightningDataModule):
         class_distribution = np.array(class_distribution)
         class_distribution = 1 / class_distribution
         class_distribution = class_distribution * len(self.class_subset) / sum(class_distribution)
-        return torch.from_numpy(class_distribution)
+        return torch.from_numpy(class_distribution).to(torch.float32)
     
     def _collate_X_y(self, batch):
         X, y = list(zip(*batch))
